Refactor ugly condition in inline_display function
Simplify the complex condition that determines when emphasis or code span
can start. The original condition was marked with a TODO comment saying
'this condition looks ugly'.

The refactored condition maintains the same logic but is more readable:
- Allows formatting at start of line (i == c)
- Allows formatting after escaped characters (*i == L'\')
- Allows formatting after whitespace (iswspace(*(i - 1)))
- Allows formatting after emphasis markup (* (i - 1) == L'*' || *(i - 1) == L'_')

This addresses issue #169 by cleaning up the code style and improving
maintainability.